rename userProject to projectContributor
